requirements.txt
# windows-curses for windows development

python-dotenv==1.0.1
requests==2.32.3
PyYAML==6.0.2
mysql-connector-python==8.3.0
dnspython==2.6.1
discord.py==2.3.1
discord.py-Console==0.1.2
aiomysql==0.2.0